moved ManagedPeersHolder interface to common package in order to break imports
