+++ integrationTests/mock/cryptoComponentsStub.go
@@ -5,9 +5,9 @@ import (
 	"sync"
 
 	crypto "github.com/multiversx/mx-chain-crypto-go"
+	"github.com/multiversx/mx-chain-go/common"
 	cryptoCommon "github.com/multiversx/mx-chain-go/common/crypto"
 	"github.com/multiversx/mx-chain-go/consensus"
-	"github.com/multiversx/mx-chain-go/keysManagement"
 	"github.com/multiversx/mx-chain-go/vm"
 )
 
@@ -29,7 +29,7 @@ type CryptoComponentsStub struct {
 	TxKeyGen                crypto.KeyGenerator
 	P2PKeyGen               crypto.KeyGenerator
 	MsgSigVerifier          vm.MessageSignVerifier
-	ManagedPeersHolderField keysManagement.ManagedPeersHolder
+	ManagedPeersHolderField common.ManagedPeersHolder
 	KeysHandlerField        consensus.KeysHandler
 	SigHandler              consensus.SigningHandler
 	mutMultiSig             sync.RWMutex
@@ -163,7 +163,7 @@ func (ccs *CryptoComponentsStub) ConsensusSigningHandler() consensus.SigningHand
 }
 
 // ManagedPeersHolder -
-func (ccs *CryptoComponentsStub) ManagedPeersHolder() keysManagement.ManagedPeersHolder {
+func (ccs *CryptoComponentsStub) ManagedPeersHolder() common.ManagedPeersHolder {
 	return ccs.ManagedPeersHolderField
 }
